Final/requirements.txt
python-dotenv~=1.0.1
PyMySQL~=1.1.1
python-telegram-bot~=21.10
